Control.Monad.Logger.Aeson.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with care.

Message-related

data Message

A Message captures a textual component and a metadata component. The metadata component is a list of SeriesElem to support tacking on arbitrary structured data to a log message.

With the OverloadedStrings extension enabled, Message values can be constructed without metadata fairly conveniently, just as if we were using Text directly:

logDebug "Some log message without metadata"

Metadata may be included in a Message via the :# constructor:

logDebug $ "Some log message with metadata" :#
  [ "bloorp" .= (42 :: Int)
  , "bonk" .= ("abc" :: Text)
  ]

The mnemonic for the :# constructor is that the # symbol is sometimes referred to as a hash, a JSON object can be thought of as a hash map, and so with :# (and enough squinting), we are cons-ing a textual message onto a JSON object. Yes, this mnemonic isn't well-typed, but hopefully it still helps!

Since: 0.1.0.0

Constructors

Text :# [SeriesElem] infixr 5

newtype SeriesElem

A single key-value pair, where the value is encoded JSON. This is a more restricted version of Series: a SeriesElem is intended to encapsulate exactly one key-value pair, whereas a Series encapsulates zero or more key-value pairs. SeriesElem values can be created via (.=) from aeson.

While a SeriesElem most often will map to a single pair, note that a Semigroup instance is available for performance's sake. The Semigroup instance is useful when multiple pairs are grouped together and then shared across multiple logging calls. In that case, the cost of combining the pairs in the group must only be paid once.

Since: 0.3.0.0

Constructors

UnsafeSeriesElem

Fields

unSeriesElem :: Series

data LoggedMessage

This type is the Haskell representation of each JSON log message produced by this library.

While we never interact with this type directly when logging messages with monad-logger-aeson, we may wish to use this type if we are parsing/processing log files generated by this library.

Since: 0.1.0.0

Constructors

LoggedMessage

Fields

loggedMessageTimestamp :: UTCTime

loggedMessageLevel :: LogLevel

loggedMessageLoc :: Maybe Loc

loggedMessageLogSource :: Maybe LogSource

loggedMessageThreadContext :: KeyMap Value

loggedMessageText :: Text

loggedMessageMeta :: KeyMap Value

threadContextStore :: Store (KeyMap Value)

Thread-safe, global Store that captures the thread context of messages.

Note that there is a bit of somewhat unavoidable name-overloading here: this binding is called threadContextStore because it stores the thread context (i.e. ThreadContext/MDC from Java land) for messages. It also just so happens that the Store type comes from the context package, which is a package providing thread-indexed storage of arbitrary context values. Please don't hate the player!

Since: 0.1.0.0

logCS :: MonadLogger m => CallStack -> LogSource -> LogLevel -> Message -> m ()

data OutputOptions

OutputOptions is for use with defaultOutputWith and enables us to configure the JSON output produced by this library.

We can get a hold of a value of this type via defaultOutputOptions.

Since: 0.1.0.0

Constructors

OutputOptions

Fields

outputAction :: LogLevel -> ByteString -> IO ()

outputIncludeThreadId :: Bool

Controls whether or not the thread ID is included in each log message's thread context.

Default: False

Since: 0.1.0.0

outputBaseThreadContext :: [Pair]

Allows for setting a "base" thread context, i.e. a set of Pair that will always be present in log messages.

If we subsequently use withThreadContext to register some thread context for our messages, if any of the keys in those Pair values overlap with the "base" thread context, then the overlapped Pair values in the "base" thread context will be overridden for the duration of the action provided to withThreadContext.

Default: mempty

Since: 0.1.0.0

defaultLogStrBS :: UTCTime -> KeyMap Value -> Loc -> LogSource -> LogLevel -> LogStr -> ByteString

defaultLogStrLBS :: UTCTime -> KeyMap Value -> Loc -> LogSource -> LogLevel -> LogStr -> ByteString

messageEncoding :: Message -> Encoding

messageSeries :: Message -> Series

LogItem-related

data LogItem

Constructors

LogItem

Fields

logItemTimestamp :: UTCTime

logItemLoc :: Loc

logItemLogSource :: LogSource

logItemLevel :: LogLevel

logItemThreadContext :: KeyMap Value

logItemMessageEncoding :: Encoding

logItemEncoding :: LogItem -> Encoding

Encoding-related

pairsEncoding :: [Pair] -> Encoding

pairsSeries :: [Pair] -> Series

levelEncoding :: LogLevel -> Encoding

locEncoding :: Loc -> Encoding

monad-logger internals

mkLoggerLoc :: SrcLoc -> Loc

Not exported from 'monad-logger', so copied here.

locFromCS :: CallStack -> Loc

Not exported from 'monad-logger', so copied here.

isDefaultLoc :: Loc -> Bool

Not exported from 'monad-logger', so copied here.

Aeson compat

data Key

data KeyMap v

A map from JSON key type Key to v.

emptyKeyMap :: KeyMap v

keyMapFromList :: [(Key, v)] -> KeyMap v

keyMapToList :: KeyMap v -> [(Key, v)]

keyMapInsert :: Key -> v -> KeyMap v -> KeyMap v

keyMapUnion :: KeyMap v -> KeyMap v -> KeyMap v